pyproject.toml
[project]
name = "awesome-generative-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.0.27",
    "langchain-openai>=0.3.6",
    "numpy>=2.2.3",
    "openai>=1.63.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
]
